float

+++ application/V2/models/CustomerGrade.php
@@ -257,12 +257,12 @@ class CustomerGradeModel extends PublicModel {
             return false;
         }
         $info=$this->where(array('id'=>$data['id']))->find();
-        $info['amount']=intval($info['amount']);
-        $info['re_purchase']=intval($info['re_purchase']);
+        $info['amount']=floatval($info['amount']);
+        $info['re_purchase']=floatval($info['re_purchase']);
 
-        $info['purchase']=intval($info['purchase']);
-        $info['income']=intval($info['income']);
-        $info['scale']=intval($info['scale']);
+        $info['purchase']=floatval($info['purchase']);
+        $info['income']=floatval($info['income']);
+        $info['scale']=floatval($info['scale']);
         if($lang=='zh'){
             $info['customer_grade']=mb_substr($info['customer_grade'],0,1).' 级';
         }else{
